forthebirds/pyproject.toml
[tool.poetry]
name = "forthebirds"
version = "0.1.1"
description = "Python package that uses the eBird API and scrapes AllAboutBirds to enable birders to find and ID more birds!"
authors = ["Lauren Deitz"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.4"
bs4 = "^0.0.1"
requests = "^2.26.0"
pytest = "^6.2.5"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"
myst-nb = "^0.13.1"
sphinx-autoapi = "^1.8.4"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
